refactor: consolidate hooks
