+++ packages/application/src/hooks/useComponents.ts
@@ -1,7 +1,28 @@
-import { useCallback, useState } from 'react';
+import type {
+  ComponentCompilerRequest,
+  ComponentCompilerResponse,
+} from '@bos-web-engine/compiler';
+import { useCallback, useEffect, useState } from 'react';
 
-export function useComponents() {
+import type { UseWebEngineParams } from '../types';
+
+interface CompilerWorker extends Omit<Worker, 'postMessage'> {
+  postMessage(compilerRequest: ComponentCompilerRequest): void;
+}
+
+export function useComponents({
+  config,
+  rootComponentPath,
+}: UseWebEngineParams) {
+  const [compiler, setCompiler] = useState<CompilerWorker | null>(null);
   const [components, setComponents] = useState<{ [key: string]: any }>({});
+  const [isCompilerInitialized, setIsCompilerInitialized] = useState(false);
+  const [isValidRootComponentPath, setIsValidRootComponentPath] =
+    useState(false);
+  const [error, setError] = useState<string | null>(null);
+  const [rootComponentSource, setRootComponentSource] = useState<string | null>(
+    null
+  );
 
   const addComponent = useCallback((componentId: string, component: any) => {
     setComponents((currentComponents) => ({
@@ -21,7 +42,20 @@ export function useComponents() {
     [components]
   );
 
-  const componentRendered = (componentId: string) =>
+  const hooks = { ...config.hooks } || {};
+  const { flags, preactVersion } = config;
+
+  useEffect(() => {
+    setIsValidRootComponentPath(
+      !!rootComponentPath &&
+        /^((([a-z\d]+[\-_])*[a-z\d]+\.)*([a-z\d]+[\-_])*[a-z\d]+)\/[\w.-]+$/gi.test(
+          rootComponentPath
+        )
+    );
+  }, [rootComponentPath]);
+
+  hooks.componentRendered = (componentId: string) => {
+    config.hooks?.componentRendered?.(componentId);
     setComponents((currentComponents) => ({
       ...currentComponents,
       [componentId]: {
@@ -29,11 +63,91 @@ export function useComponents() {
         renderCount: currentComponents?.[componentId]?.renderCount + 1 || 0,
       },
     }));
+  };
+
+  useEffect(() => {
+    if (!rootComponentPath || !isValidRootComponentPath) {
+      return;
+    }
+
+    if (!compiler) {
+      const worker = new Worker(
+        new URL('../workers/compiler.js', import.meta.url)
+      );
+      const initPayload: ComponentCompilerRequest = {
+        action: 'init',
+        localFetchUrl: flags?.bosLoaderUrl,
+        preactVersion,
+      };
+      worker.postMessage(initPayload);
+      setCompiler(worker);
+    } else if (!isCompilerInitialized) {
+      setIsCompilerInitialized(true);
+
+      compiler.onmessage = ({
+        data,
+      }: MessageEvent<ComponentCompilerResponse>) => {
+        const {
+          componentId,
+          componentSource,
+          error: loadError,
+          importedModules,
+        } = data;
+
+        if (loadError) {
+          setError(loadError.message);
+          return;
+        }
+
+        hooks?.containerSourceCompiled?.(data);
+
+        // set the Preact import maps
+        // TODO find a better place for this
+        importedModules.set('preact', `https://esm.sh/preact@${preactVersion}`);
+        importedModules.set(
+          'preact/',
+          `https://esm.sh/preact@${preactVersion}/`
+        );
+
+        const component = {
+          ...components[componentId],
+          componentId,
+          componentSource,
+          moduleImports: importedModules,
+        };
+
+        if (!rootComponentSource && componentId === rootComponentPath) {
+          setRootComponentSource(componentId);
+        }
+
+        addComponent(componentId, component);
+      };
+
+      compiler.postMessage({
+        action: 'execute',
+        componentId: rootComponentPath,
+      });
+    }
+  }, [
+    rootComponentPath,
+    rootComponentSource,
+    compiler,
+    addComponent,
+    components,
+    isCompilerInitialized,
+    error,
+    isValidRootComponentPath,
+    flags?.bosLoaderUrl,
+    preactVersion,
+    hooks,
+  ]);
 
   return {
     addComponent,
+    compiler,
     components,
-    componentRendered,
+    error,
+    hooks,
     getComponentRenderCount,
     setComponents,
   };
